requirements.txt
asgiref==3.8.1
boto3==1.35.54
botocore==1.35.54
crispy-bootstrap4==2024.10
dj-database-url==2.3.0
Django==4.2
django-allauth==65.1.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
phonenumberslite==8.13.48
pillow==11.0.0
psycopg2==2.9.10
s3transfer==0.10.3
sqlparse==0.5.1
